Updated sendToViewers module for networked access and IRIS container

+++ utils/sendToViewers.js
@@ -20,35 +20,56 @@ module.exports = function(global, subscripts) {
     if (subscripts.length === 1) subs = "'" + subscripts[0] + "'";
     jsdbCommand = jsdbCommand + '.$(' + subs + ')';
   }
-  jsdbCommand = jsdbCommand + '.getDocument(true)';  
+  jsdbCommand = jsdbCommand + '.getDocument(true)';
 
-  var status = this.db.dbx.function('zwr^qewdInterface', ref);
-  var prefix = '/opt/qewd/mapped/';
-  if (fs.existsSync('/ISC/dur')) {
-    prefix = '';
+  var stream = false;
+  if (this.userDefined.config.database.params && this.userDefined.config.database.params.host) {
+    stream = true;
   }
-  var filePath = prefix + 'go-' + process.pid + '.txt';
-  if (fs.existsSync(filePath)) {
-    var text = fs.readFileSync(filePath, 'utf8');
-    fs.removeSync(filePath);
 
-    var json = this.documentStore.use(global, subscripts).getDocument(true);
-
-    var activeSessions = this.sessions.active();
-    activeSessions.forEach(function(qewdSession) {
-      console.log('session: ' + qewdSession.id);
-      if (qewdSession.application === 'jsdb-viewer' && qewdSession.socketId) {
-        process.send({
-          socketId: qewdSession.socketId,
-          type: 'view-test',
-          globalRef: ref,
-          globalContent: text,
-          jsdbDocRef: jsdbDocRef,
-          jsdbCommand: jsdbCommand,
-          jsonContent: json
-        });
-      }
-    });
+  var status;
+  if (stream) {
+    var status = this.db.dbx.function('zwr^qewdInterface', ref, 'true');
   }
+  else {
+    status = this.db.dbx.function('zwr^qewdInterface', ref);
+  }
+
+  var text = '';
+  if (stream) {
+    text = status;
+  }
+  else {
+    var prefix = '/opt/qewd/mapped/';
+    if (fs.existsSync('/usr/irissys/mgr/user')) {
+      prefix = '/usr/irissys/mgr/user/';
+    }
+    if (fs.existsSync('/ISC/dur')) {
+      prefix = '';
+    }
+    var filePath = prefix + status;
+    if (fs.existsSync(filePath)) {
+      text = fs.readFileSync(filePath, 'utf8');
+      fs.removeSync(filePath);
+    }
+  }
+
+  var json = this.documentStore.use(global, subscripts).getDocument(true);
+
+  var activeSessions = this.sessions.active();
+  activeSessions.forEach(function(qewdSession) {
+    console.log('session: ' + qewdSession.id);
+    if (qewdSession.application === 'jsdb-viewer' && qewdSession.socketId) {
+      process.send({
+        socketId: qewdSession.socketId,
+        type: 'view-test',
+        globalRef: ref,
+        globalContent: text,
+        jsdbDocRef: jsdbDocRef,
+        jsdbCommand: jsdbCommand,
+        jsonContent: json
+      });
+    }
+  });
 
 };
